% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factor_Order.R
\name{Factor_Order}
\alias{Factor_Order}
\title{Compute levels of factor}
\usage{
Factor_Order(x)
}
\arguments{
\item{x}{is a factor}
}
\value{
Return levels to the order in which they appear in the data,
An error is raised if the input is not a factor
}
\description{
This function computes the levels of a given factor by returning the unique ordering found in the dataset.
}
\examples{
Factor_Order(c("high", "med", "low"))
}
