% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factor_Reorder.R
\name{Factor_Reorder}
\alias{Factor_Reorder}
\title{Reorder a factor X in descending order}
\usage{
Factor_Reorder
}
\arguments{
\item{x}{Factor for reordering}
}
\value{
Returns the same factor x with levels ordered in descending fashion
Note: The input must be a factor or an exception will be thrown.
}
\description{
This function takes in a factor x and reorder its levels in descending order. We appeal to the desc function from dplyr
}
\examples{
Factor_Reorder(c("high", "med", "low"))
}
