% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isFactor.R
\name{isFactor}
\alias{isFactor}
\title{Checks if input is a factor or a string}
\usage{
isFactor
}
\arguments{
\item{x}{}
}
\value{
boolean
}
\description{
A function that checks if a an input x can be represented as factor or not and returns
a boolean value.
}
\examples{
isFactor(c("high", "medium", "low"))
}
