% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_textFile.R
\name{Read_textFile}
\alias{Read_textFile}
\title{Read a text file  into a dataframe}
\usage{
Read_textFile()
}
\arguments{
\item{file}{name of the file you'd like to read in}
}
\description{
Read a text file  into a dataframe
}
