% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Write_textFile.R
\name{Write_textFile}
\alias{Write_textFile}
\title{Write a dataframe  into a textfile}
\usage{
Write_textFile()
}
\arguments{
\item{dataframe}{A dataframe}

\item{name}{The name you would like the file to be saved under}

\item{delim}{The delimiter you would like to use to write the dataframe}
}
\description{
Write a dataframe  into a textfile
}
